backend/requirements.txt
# ============================================================
# AI Knowledge Navigator — Python Dependencies
# Install: pip install -r requirements.txt
# ============================================================

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
motor==3.4.0           # Async MongoDB driver
pymongo==4.7.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# AI — OpenAI
openai==1.30.1

# AI — Google Gemini (uncomment if using Gemini)
# google-generativeai==0.5.4

# PDF Processing
pdfplumber==0.11.0
PyPDF2==3.0.1

# Vector Database
faiss-cpu==1.8.0

# Numpy
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
